Play audio from cache; Fix interruptions in ASR (#12)
* Play audio

This change enables TTS to play audio. The audio is sent as part of SPEAK
event (`audioURL` parameter). If the resource is not available, it falls back
to normal synthesis (`utterance` parameter).

Additionally, for streamed content, the `cache` parameter containing a URL
can be provided with the SPEAK event. This URL will then be first called to
check whether cache exists and then, (true) make another request to get
an audio file to play or (false) fallback to normal TTS.

* Fix interruptions in ASR while resources are being respawned

This bugfix removes respawning of ASR and TTS machines. Instead, the
parent machine (`speechstate.ts`) generates new tokens and
communicates them to ASR and TTS via NEW_TOKEN event.

ASR doesn’t need to be pre-initiated, so only the context gets updated
for a new token. New token is then used for a new ASR request.

TTS *needs* to be pre-initiated, therefore on a new token it generates
new instances of Web Speech API synthesiser and utterance. They are
then used for the next TTS request.

Appropriate tests are added which attempt to renew tokens while the
system is recognising or speaking. Additionally, a new setting,
`newTokenInterval` is added to configure how often new tokens will be
requested (mainly for testing purposes), defaulting to 30
seconds. Unfortunately, the tests are not exactly fair because the old
token doesn’t expire when the new one is released, but I tried to
make sure that the new token is actually used.