Stop streaming after timeout for next chunk (#16)
* Stop streaming after timeout for next chunk

This change introduces STREAMING_TIMEOUT (set to 10 seconds). When the
timeout is reached, TTS moves to a idle state and throws a
SPEAK_COMPLETE event, like it normally does after
speaking. Additionally, an error is logged in browser console.

This change also enables logging the utterances which are played from
a cache.